change execute and executeAsync to return the last event instead of status since it will almost always be 0
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@15973 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ libs/esl/java/org/freeswitch/esl/eslJNI.java
@@ -44,8 +44,8 @@ class eslJNI {
   public final static native long ESLconnection_recvEventTimed(long jarg1, ESLconnection jarg1_, int jarg2);
   public final static native long ESLconnection_filter(long jarg1, ESLconnection jarg1_, String jarg2, String jarg3);
   public final static native int ESLconnection_events(long jarg1, ESLconnection jarg1_, String jarg2, String jarg3);
-  public final static native int ESLconnection_execute(long jarg1, ESLconnection jarg1_, String jarg2, String jarg3, String jarg4);
-  public final static native int ESLconnection_executeAsync(long jarg1, ESLconnection jarg1_, String jarg2, String jarg3, String jarg4);
+  public final static native long ESLconnection_execute(long jarg1, ESLconnection jarg1_, String jarg2, String jarg3, String jarg4);
+  public final static native long ESLconnection_executeAsync(long jarg1, ESLconnection jarg1_, String jarg2, String jarg3, String jarg4);
   public final static native int ESLconnection_setAsyncExecute(long jarg1, ESLconnection jarg1_, String jarg2);
   public final static native int ESLconnection_setEventLock(long jarg1, ESLconnection jarg1_, String jarg2);
   public final static native int ESLconnection_disconnect(long jarg1, ESLconnection jarg1_);
